+++ CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## v2.16.1 - 2023-07-27
+
 ### Fixed
 
 - Allow empty environment variables for the directory
 - Better health checks in directory
+- Better handling of persistence file at load time of directory
 
 ## v2.16.0 - 2023-07-24